Resolving defect regarding successful new registration where we now automatically perform an Auth Key Backup on behalf of the user.
We spawn a child process to execute auth-keys-backup.js The resolution to this bug was to correct a typo to start executing the correctly named nodejs script. We were previously trying to reference auth-keys-backup-script.js which is a file that did NOT exist. Dockerfile and correct script filename were both referencing 'auth-keys-backup.js' which was correct, source code however was referencing the outdated script name of 'auth-keys-backup-script.js'